+++ osquery/runtime/runtime_test.go
@@ -1,4 +1,4 @@
-package osquery
+package runtime
 
 import (
 	"bufio"
